Reorganize pyanaconda.modules.boss
Move classes for the kickstart dispatcher to the kickstart_manager
package in pyanaconda.modules.boss. Also improve the install_manager
package a little.
