+++ pyanaconda/modules/boss/boss.py
@@ -24,8 +24,7 @@ from pyanaconda.modules.boss.boss_interface import AnacondaBossInterface
 from pyanaconda.modules.common.base import BaseModule
 from pyanaconda.dbus.constants import DBUS_BOSS_NAME, DBUS_BOSS_PATH, DBUS_BOSS_INSTALLATION_PATH
 from pyanaconda.modules.boss.module_manager import ModuleManager
-from pyanaconda.modules.boss.install_manager.installation_interface import InstallationInterface
-from pyanaconda.modules.boss.install_manager.install_manager import InstallManager
+from pyanaconda.modules.boss.install_manager import InstallManager, InstallationInterface
 from pyanaconda.modules.boss.kickstart_manager import KickstartManager
 
 from pyanaconda.anaconda_loggers import get_module_logger
