+++ pyanaconda/modules/boss/install_manager/install_interface.py
