+++ pyanaconda/modules/boss/kickstart_manager/element.py
