+++ pyanaconda/modules/boss/kickstart_manager/parser.py
@@ -20,7 +20,9 @@ __all__ = ["SplitKickstartParser", "VALID_SECTIONS_ANACONDA"]
 
 from pykickstart.parser import KickstartParser
 from pykickstart.sections import Section
-from pyanaconda.kickstart_dispatcher.element import TrackedKickstartElements, KickstartElement
+
+from pyanaconda.modules.boss.kickstart_manager.element import KickstartElement,\
+    TrackedKickstartElements
 
 VALID_SECTIONS_ANACONDA = ["%pre", "%pre-install", "%post", "%onerror", "%traceback",
                            "%packages", "%addon", "%anaconda"]
